Don't parse dist.ini to get distribution name and version
There are Dist::Zilla plugins like AutoVersion or Git::NextVersion which
make it possible to omit the version from dist.ini. Because of this, there
were situations where the paths constructed from this (potentially missing)
information in dist.ini pointed to inexistent directories. These issues
can be avoided by passing the --in parameter to 'dzil build', which makes
it put the output into a known directory which can then be used.

For this known directory, I've chosen .build/debian-build since this will
also be removed when doing a 'dzil clean'.