Let dh_dist_zilla_origtar run "dzil clean" first by default
Avoids to get artifacts from previous, potentially manual builds
included in the tar ball.

Use option --no-clean to get previous behaviour.

A short note about just bumping the micro version and not the minor or
even major version:

I'm considering this to be a bug fix, not a new (potentially
non-backwards-compatible) feature, because from the documentation, the
user does not expect to get build artifacts to be included in the
.orig.tar.xz tar ball.

And yes, I ran into that myself only noticed it when dpkg was arguing
about no more present files and listed ".build" among others.